Work on the Game Tutorial
This was the hardest part of development so far, getting the link between the story, the locations and the UI linked up correctly.

I'm still not 100% sure this is the right connection layout, but it's working and so we'll go with it.

Notice that all story elements and locations are added to a load function, with the class added against a key, and then the key can be used to initialise the class into an object as and when each element/location is needed.