Items and Inventory; Logs
This commit focussed on two things. Firstly, items and inventory management for the player, which will be used both for things they can sell, and for quest items. Right now there will be no management of equippable items to keep things lean, but this could be a future addition.

Secondly, logs. Logs store the messages that the user gets sent. These then load when a user reloads the page or comes back to the game after an absence. Logs are important as they will bring the user back into the game quickly and they can see their past actions up until a limit. Right now this limit will be set between 50-150 log entries.

This commit also downgrades jQuery from 2.2.0 to 2.1.4 to fix compatibility issues with Foundation 5's reveal modal (used for the travel modal that has been added) and adds the Tipped library.